src/backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
sqlalchemy==2.0.34
aiohttp==3.10.5
redis==5.2.0
aioredis==2.0.1
pydantic==2.10.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
pytest==8.0.0
httpx==0.27.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pytest-asyncio==0.23.5
pytest-cov==4.1.0
black==24.1.1
isort==5.13.2
flake8==7.0.0 